// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type LandingPageDocumentDataSlicesSlice = HeroSlice | StoryStatsSlice | SportsGridSlice | TrainingLocationsSlice | NewsShowcaseSlice | JoinBandSlice | TextBlockSlice | ImageHighlightSlice | CallToActionSlice

/**
 * Content for Landing page documents
 */
interface LandingPageDocumentData {
	/**
	 * Page kind field in *Landing page*
	 *
	 * - **Field Type**: Select
	 * - **Placeholder**: *None*
	 * - **API ID Path**: landing_page.page_kind
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/fields/select
	 */
	page_kind: prismic.SelectField<"home" | "training" | "calendar" | "contact" | "sitemap" | "other">;
	
	/**
	 * Featured image field in *Landing page*
	 *
	 * - **Field Type**: Image
	 * - **Placeholder**: *None*
	 * - **API ID Path**: landing_page.featured_image
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/fields/image
	 */
	featured_image: prismic.ImageField<never>;
	
	/**
	 * Translation group id field in *Landing page*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: Polylang pll_* group slug from WordPress
	 * - **API ID Path**: landing_page.translation_group_id
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/fields/text
	 */
	translation_group_id: prismic.KeyTextField;
	
	/**
	 * Slice Zone field in *Landing page*
	 *
	 * - **Field Type**: Slice Zone
	 * - **Placeholder**: *None*
	 * - **API ID Path**: landing_page.slices[]
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/slices
	 */
	slices: prismic.SliceZone<LandingPageDocumentDataSlicesSlice>;/**
	 * Meta Title field in *Landing page*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: A title of the page used for social media and search engines
	 * - **API ID Path**: landing_page.meta_title
	 * - **Tab**: SEO & Metadata
	 * - **Documentation**: https://prismic.io/docs/fields/text
	 */
	meta_title: prismic.KeyTextField;
	
	/**
	 * Meta Description field in *Landing page*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: A brief summary of the page
	 * - **API ID Path**: landing_page.meta_description
	 * - **Tab**: SEO & Metadata
	 * - **Documentation**: https://prismic.io/docs/fields/text
	 */
	meta_description: prismic.KeyTextField;
	
	/**
	 * Meta Image field in *Landing page*
	 *
	 * - **Field Type**: Image
	 * - **Placeholder**: *None*
	 * - **API ID Path**: landing_page.meta_image
	 * - **Tab**: SEO & Metadata
	 * - **Documentation**: https://prismic.io/docs/fields/image
	 */
	meta_image: prismic.ImageField<never>;
}

/**
 * Landing page document from Prismic
 *
 * - **API ID**: `landing_page`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type LandingPageDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<LandingPageDocumentData>, "landing_page", Lang>;